refactor(hyperswitch_domain_models): move incorrectly placed `Conversion` implementation to correct module
