+++ crates/hyperswitch_domain_models/src/payments/payment_intent.rs
@@ -3,16 +3,30 @@ use common_utils::{
     consts::{PAYMENTS_LIST_MAX_LIMIT_V1, PAYMENTS_LIST_MAX_LIMIT_V2},
     crypto::Encryptable,
     encryption::Encryption,
+    errors::{CustomResult, ValidationError},
     id_type,
     pii::{self, Email},
-    types::{keymanager::KeyManagerState, MinorUnit},
+    type_name,
+    types::{
+        keymanager::{self, KeyManagerState},
+        MinorUnit,
+    },
 };
-use masking::{Deserialize, Secret};
+use diesel_models::{
+    PaymentIntent as DieselPaymentIntent, PaymentIntentNew as DieselPaymentIntentNew,
+};
+use error_stack::ResultExt;
+use masking::{Deserialize, PeekInterface, Secret};
 use serde::Serialize;
 use time::PrimitiveDateTime;
 
 use super::{payment_attempt::PaymentAttempt, PaymentIntent};
-use crate::{errors, merchant_key_store::MerchantKeyStore, RemoteStorageObject};
+use crate::{
+    behaviour, errors,
+    merchant_key_store::MerchantKeyStore,
+    type_encryption::{crypto_operation, AsyncLift, CryptoOperation},
+    RemoteStorageObject,
+};
 #[async_trait::async_trait]
 pub trait PaymentIntentInterface {
     async fn update_payment_intent(
@@ -1561,5 +1575,427 @@ where
         } else {
             Ok(payment_intent_constraints)
         }
+    }
+}
+
+#[cfg(all(feature = "v2", feature = "payment_v2"))]
+#[async_trait::async_trait]
+impl behaviour::Conversion for PaymentIntent {
+    type DstType = DieselPaymentIntent;
+    type NewDstType = DieselPaymentIntentNew;
+
+    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
+        Ok(DieselPaymentIntent {
+            merchant_id: self.merchant_id,
+            status: self.status,
+            amount: self.amount,
+            currency: self.currency,
+            amount_captured: self.amount_captured,
+            customer_id: self.customer_id,
+            description: self.description,
+            return_url: self.return_url,
+            metadata: self.metadata,
+            statement_descriptor_name: self.statement_descriptor_name,
+            created_at: self.created_at,
+            modified_at: self.modified_at,
+            last_synced: self.last_synced,
+            setup_future_usage: self.setup_future_usage,
+            off_session: self.off_session,
+            client_secret: self.client_secret,
+            active_attempt_id: self.active_attempt.get_id(),
+            order_details: self.order_details,
+            allowed_payment_method_types: self.allowed_payment_method_types,
+            connector_metadata: self.connector_metadata,
+            feature_metadata: self.feature_metadata,
+            attempt_count: self.attempt_count,
+            profile_id: self.profile_id,
+            frm_merchant_decision: self.frm_merchant_decision,
+            payment_link_id: self.payment_link_id,
+            payment_confirm_source: self.payment_confirm_source,
+            updated_by: self.updated_by,
+            surcharge_applicable: self.surcharge_applicable,
+            request_incremental_authorization: self.request_incremental_authorization,
+            authorization_count: self.authorization_count,
+            session_expiry: self.session_expiry,
+            request_external_three_ds_authentication: self.request_external_three_ds_authentication,
+            charges: self.charges,
+            frm_metadata: self.frm_metadata,
+            customer_details: self.customer_details.map(Encryption::from),
+            billing_address: self.billing_address.map(Encryption::from),
+            merchant_order_reference_id: self.merchant_order_reference_id,
+            shipping_address: self.shipping_address.map(Encryption::from),
+            is_payment_processor_token_flow: self.is_payment_processor_token_flow,
+            capture_method: self.capture_method,
+            id: self.id,
+            authentication_type: self.authentication_type,
+            amount_to_capture: self.amount_to_capture,
+            prerouting_algorithm: self.prerouting_algorithm,
+            merchant_reference_id: self.merchant_reference_id,
+            surcharge_amount: self.surcharge_amount,
+            tax_on_surcharge: self.tax_on_surcharge,
+            organization_id: self.organization_id,
+            shipping_cost: self.shipping_cost,
+            tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
+        })
+    }
+    async fn convert_back(
+        state: &KeyManagerState,
+        storage_model: Self::DstType,
+        key: &Secret<Vec<u8>>,
+        key_manager_identifier: keymanager::Identifier,
+    ) -> CustomResult<Self, ValidationError>
+    where
+        Self: Sized,
+    {
+        async {
+            let inner_decrypt = |inner| async {
+                crypto_operation(
+                    state,
+                    type_name!(Self::DstType),
+                    CryptoOperation::DecryptOptional(inner),
+                    key_manager_identifier.clone(),
+                    key.peek(),
+                )
+                .await
+                .and_then(|val| val.try_into_optionaloperation())
+            };
+            Ok::<Self, error_stack::Report<common_utils::errors::CryptoError>>(Self {
+                merchant_id: storage_model.merchant_id,
+                status: storage_model.status,
+                amount: storage_model.amount,
+                currency: storage_model.currency,
+                amount_captured: storage_model.amount_captured,
+                customer_id: storage_model.customer_id,
+                description: storage_model.description,
+                return_url: storage_model.return_url,
+                metadata: storage_model.metadata,
+                statement_descriptor_name: storage_model.statement_descriptor_name,
+                created_at: storage_model.created_at,
+                modified_at: storage_model.modified_at,
+                last_synced: storage_model.last_synced,
+                setup_future_usage: storage_model.setup_future_usage,
+                off_session: storage_model.off_session,
+                client_secret: storage_model.client_secret,
+                active_attempt: RemoteStorageObject::ForeignID(storage_model.active_attempt_id),
+                order_details: storage_model.order_details,
+                allowed_payment_method_types: storage_model.allowed_payment_method_types,
+                connector_metadata: storage_model.connector_metadata,
+                feature_metadata: storage_model.feature_metadata,
+                attempt_count: storage_model.attempt_count,
+                profile_id: storage_model.profile_id,
+                frm_merchant_decision: storage_model.frm_merchant_decision,
+                payment_link_id: storage_model.payment_link_id,
+                payment_confirm_source: storage_model.payment_confirm_source,
+                updated_by: storage_model.updated_by,
+                surcharge_applicable: storage_model.surcharge_applicable,
+                request_incremental_authorization: storage_model.request_incremental_authorization,
+                authorization_count: storage_model.authorization_count,
+                session_expiry: storage_model.session_expiry,
+                request_external_three_ds_authentication: storage_model
+                    .request_external_three_ds_authentication,
+                charges: storage_model.charges,
+                frm_metadata: storage_model.frm_metadata,
+                customer_details: storage_model
+                    .customer_details
+                    .async_lift(inner_decrypt)
+                    .await?,
+                billing_address: storage_model
+                    .billing_address
+                    .async_lift(inner_decrypt)
+                    .await?,
+                merchant_order_reference_id: storage_model.merchant_order_reference_id,
+                shipping_address: storage_model
+                    .shipping_address
+                    .async_lift(inner_decrypt)
+                    .await?,
+                is_payment_processor_token_flow: storage_model.is_payment_processor_token_flow,
+                capture_method: storage_model.capture_method,
+                id: storage_model.id,
+                merchant_reference_id: storage_model.merchant_reference_id,
+                organization_id: storage_model.organization_id,
+                authentication_type: storage_model.authentication_type,
+                amount_to_capture: storage_model.amount_to_capture,
+                prerouting_algorithm: storage_model.prerouting_algorithm,
+                surcharge_amount: storage_model.surcharge_amount,
+                tax_on_surcharge: storage_model.tax_on_surcharge,
+                shipping_cost: storage_model.shipping_cost,
+                tax_details: storage_model.tax_details,
+                skip_external_tax_calculation: storage_model.skip_external_tax_calculation,
+            })
+        }
+        .await
+        .change_context(ValidationError::InvalidValue {
+            message: "Failed while decrypting payment intent".to_string(),
+        })
+    }
+
+    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
+        Ok(DieselPaymentIntentNew {
+            merchant_id: self.merchant_id,
+            status: self.status,
+            amount: self.amount,
+            currency: self.currency,
+            amount_captured: self.amount_captured,
+            customer_id: self.customer_id,
+            description: self.description,
+            return_url: self.return_url,
+            metadata: self.metadata,
+            statement_descriptor_name: self.statement_descriptor_name,
+            created_at: self.created_at,
+            modified_at: self.modified_at,
+            last_synced: self.last_synced,
+            setup_future_usage: self.setup_future_usage,
+            off_session: self.off_session,
+            client_secret: self.client_secret,
+            active_attempt_id: self.active_attempt.get_id(),
+            order_details: self.order_details,
+            allowed_payment_method_types: self.allowed_payment_method_types,
+            connector_metadata: self.connector_metadata,
+            feature_metadata: self.feature_metadata,
+            attempt_count: self.attempt_count,
+            profile_id: self.profile_id,
+            frm_merchant_decision: self.frm_merchant_decision,
+            payment_link_id: self.payment_link_id,
+            payment_confirm_source: self.payment_confirm_source,
+            updated_by: self.updated_by,
+            surcharge_applicable: self.surcharge_applicable,
+            request_incremental_authorization: self.request_incremental_authorization,
+            authorization_count: self.authorization_count,
+            session_expiry: self.session_expiry,
+            request_external_three_ds_authentication: self.request_external_three_ds_authentication,
+            charges: self.charges,
+            frm_metadata: self.frm_metadata,
+            customer_details: self.customer_details.map(Encryption::from),
+            billing_address: self.billing_address.map(Encryption::from),
+            merchant_order_reference_id: self.merchant_order_reference_id,
+            shipping_address: self.shipping_address.map(Encryption::from),
+            is_payment_processor_token_flow: self.is_payment_processor_token_flow,
+            capture_method: self.capture_method,
+            id: self.id,
+            merchant_reference_id: self.merchant_reference_id,
+            authentication_type: self.authentication_type,
+            amount_to_capture: self.amount_to_capture,
+            prerouting_algorithm: self.prerouting_algorithm,
+            surcharge_amount: self.surcharge_amount,
+            tax_on_surcharge: self.tax_on_surcharge,
+            organization_id: self.organization_id,
+            shipping_cost: self.shipping_cost,
+            tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
+        })
+    }
+}
+
+#[cfg(all(any(feature = "v1", feature = "v2"), not(feature = "payment_v2")))]
+#[async_trait::async_trait]
+impl behaviour::Conversion for PaymentIntent {
+    type DstType = DieselPaymentIntent;
+    type NewDstType = DieselPaymentIntentNew;
+
+    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
+        Ok(DieselPaymentIntent {
+            payment_id: self.payment_id,
+            merchant_id: self.merchant_id,
+            status: self.status,
+            amount: self.amount,
+            currency: self.currency,
+            amount_captured: self.amount_captured,
+            customer_id: self.customer_id,
+            description: self.description,
+            return_url: self.return_url,
+            metadata: self.metadata,
+            connector_id: self.connector_id,
+            shipping_address_id: self.shipping_address_id,
+            billing_address_id: self.billing_address_id,
+            statement_descriptor_name: self.statement_descriptor_name,
+            statement_descriptor_suffix: self.statement_descriptor_suffix,
+            created_at: self.created_at,
+            modified_at: self.modified_at,
+            last_synced: self.last_synced,
+            setup_future_usage: self.setup_future_usage,
+            off_session: self.off_session,
+            client_secret: self.client_secret,
+            active_attempt_id: self.active_attempt.get_id(),
+            business_country: self.business_country,
+            business_label: self.business_label,
+            order_details: self.order_details,
+            allowed_payment_method_types: self.allowed_payment_method_types,
+            connector_metadata: self.connector_metadata,
+            feature_metadata: self.feature_metadata,
+            attempt_count: self.attempt_count,
+            profile_id: self.profile_id,
+            merchant_decision: self.merchant_decision,
+            payment_link_id: self.payment_link_id,
+            payment_confirm_source: self.payment_confirm_source,
+            updated_by: self.updated_by,
+            surcharge_applicable: self.surcharge_applicable,
+            request_incremental_authorization: self.request_incremental_authorization,
+            incremental_authorization_allowed: self.incremental_authorization_allowed,
+            authorization_count: self.authorization_count,
+            fingerprint_id: self.fingerprint_id,
+            session_expiry: self.session_expiry,
+            request_external_three_ds_authentication: self.request_external_three_ds_authentication,
+            charges: self.charges,
+            frm_metadata: self.frm_metadata,
+            customer_details: self.customer_details.map(Encryption::from),
+            billing_details: self.billing_details.map(Encryption::from),
+            merchant_order_reference_id: self.merchant_order_reference_id,
+            shipping_details: self.shipping_details.map(Encryption::from),
+            is_payment_processor_token_flow: self.is_payment_processor_token_flow,
+            organization_id: self.organization_id,
+            shipping_cost: self.shipping_cost,
+            tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
+        })
+    }
+
+    async fn convert_back(
+        state: &KeyManagerState,
+        storage_model: Self::DstType,
+        key: &Secret<Vec<u8>>,
+        key_manager_identifier: keymanager::Identifier,
+    ) -> CustomResult<Self, ValidationError>
+    where
+        Self: Sized,
+    {
+        async {
+            let inner_decrypt = |inner| async {
+                crypto_operation(
+                    state,
+                    type_name!(Self::DstType),
+                    CryptoOperation::DecryptOptional(inner),
+                    key_manager_identifier.clone(),
+                    key.peek(),
+                )
+                .await
+                .and_then(|val| val.try_into_optionaloperation())
+            };
+            Ok::<Self, error_stack::Report<common_utils::errors::CryptoError>>(Self {
+                payment_id: storage_model.payment_id,
+                merchant_id: storage_model.merchant_id,
+                status: storage_model.status,
+                amount: storage_model.amount,
+                currency: storage_model.currency,
+                amount_captured: storage_model.amount_captured,
+                customer_id: storage_model.customer_id,
+                description: storage_model.description,
+                return_url: storage_model.return_url,
+                metadata: storage_model.metadata,
+                connector_id: storage_model.connector_id,
+                shipping_address_id: storage_model.shipping_address_id,
+                billing_address_id: storage_model.billing_address_id,
+                statement_descriptor_name: storage_model.statement_descriptor_name,
+                statement_descriptor_suffix: storage_model.statement_descriptor_suffix,
+                created_at: storage_model.created_at,
+                modified_at: storage_model.modified_at,
+                last_synced: storage_model.last_synced,
+                setup_future_usage: storage_model.setup_future_usage,
+                off_session: storage_model.off_session,
+                client_secret: storage_model.client_secret,
+                active_attempt: RemoteStorageObject::ForeignID(storage_model.active_attempt_id),
+                business_country: storage_model.business_country,
+                business_label: storage_model.business_label,
+                order_details: storage_model.order_details,
+                allowed_payment_method_types: storage_model.allowed_payment_method_types,
+                connector_metadata: storage_model.connector_metadata,
+                feature_metadata: storage_model.feature_metadata,
+                attempt_count: storage_model.attempt_count,
+                profile_id: storage_model.profile_id,
+                merchant_decision: storage_model.merchant_decision,
+                payment_link_id: storage_model.payment_link_id,
+                payment_confirm_source: storage_model.payment_confirm_source,
+                updated_by: storage_model.updated_by,
+                surcharge_applicable: storage_model.surcharge_applicable,
+                request_incremental_authorization: storage_model.request_incremental_authorization,
+                incremental_authorization_allowed: storage_model.incremental_authorization_allowed,
+                authorization_count: storage_model.authorization_count,
+                fingerprint_id: storage_model.fingerprint_id,
+                session_expiry: storage_model.session_expiry,
+                request_external_three_ds_authentication: storage_model
+                    .request_external_three_ds_authentication,
+                charges: storage_model.charges,
+                frm_metadata: storage_model.frm_metadata,
+                shipping_cost: storage_model.shipping_cost,
+                tax_details: storage_model.tax_details,
+                customer_details: storage_model
+                    .customer_details
+                    .async_lift(inner_decrypt)
+                    .await?,
+                billing_details: storage_model
+                    .billing_details
+                    .async_lift(inner_decrypt)
+                    .await?,
+                merchant_order_reference_id: storage_model.merchant_order_reference_id,
+                shipping_details: storage_model
+                    .shipping_details
+                    .async_lift(inner_decrypt)
+                    .await?,
+                is_payment_processor_token_flow: storage_model.is_payment_processor_token_flow,
+                organization_id: storage_model.organization_id,
+                skip_external_tax_calculation: storage_model.skip_external_tax_calculation,
+            })
+        }
+        .await
+        .change_context(ValidationError::InvalidValue {
+            message: "Failed while decrypting payment intent".to_string(),
+        })
+    }
+
+    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
+        Ok(DieselPaymentIntentNew {
+            payment_id: self.payment_id,
+            merchant_id: self.merchant_id,
+            status: self.status,
+            amount: self.amount,
+            currency: self.currency,
+            amount_captured: self.amount_captured,
+            customer_id: self.customer_id,
+            description: self.description,
+            return_url: self.return_url,
+            metadata: self.metadata,
+            connector_id: self.connector_id,
+            shipping_address_id: self.shipping_address_id,
+            billing_address_id: self.billing_address_id,
+            statement_descriptor_name: self.statement_descriptor_name,
+            statement_descriptor_suffix: self.statement_descriptor_suffix,
+            created_at: self.created_at,
+            modified_at: self.modified_at,
+            last_synced: self.last_synced,
+            setup_future_usage: self.setup_future_usage,
+            off_session: self.off_session,
+            client_secret: self.client_secret,
+            active_attempt_id: self.active_attempt.get_id(),
+            business_country: self.business_country,
+            business_label: self.business_label,
+            order_details: self.order_details,
+            allowed_payment_method_types: self.allowed_payment_method_types,
+            connector_metadata: self.connector_metadata,
+            feature_metadata: self.feature_metadata,
+            attempt_count: self.attempt_count,
+            profile_id: self.profile_id,
+            merchant_decision: self.merchant_decision,
+            payment_link_id: self.payment_link_id,
+            payment_confirm_source: self.payment_confirm_source,
+            updated_by: self.updated_by,
+            surcharge_applicable: self.surcharge_applicable,
+            request_incremental_authorization: self.request_incremental_authorization,
+            incremental_authorization_allowed: self.incremental_authorization_allowed,
+            authorization_count: self.authorization_count,
+            fingerprint_id: self.fingerprint_id,
+            session_expiry: self.session_expiry,
+            request_external_three_ds_authentication: self.request_external_three_ds_authentication,
+            charges: self.charges,
+            frm_metadata: self.frm_metadata,
+            customer_details: self.customer_details.map(Encryption::from),
+            billing_details: self.billing_details.map(Encryption::from),
+            merchant_order_reference_id: self.merchant_order_reference_id,
+            shipping_details: self.shipping_details.map(Encryption::from),
+            is_payment_processor_token_flow: self.is_payment_processor_token_flow,
+            organization_id: self.organization_id,
+            shipping_cost: self.shipping_cost,
+            tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
+        })
     }
 }
